src/test/java/com/github/tatanka27/voting/web/controller/restaurant/UserRestaurantControllerTest.java
package com.github.tatanka27.voting.web.controller.restaurant;

import org.junit.jupiter.api.Test;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithUserDetails;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import com.github.tatanka27.voting.web.controller.AbstractControllerTest;

import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static com.github.tatanka27.voting.data.RestaurantTestData.RESTAURANT_MATCHER;
import static com.github.tatanka27.voting.data.RestaurantTestData.restaurants;
import static com.github.tatanka27.voting.data.UserTestData.USER_MAIL;
import static com.github.tatanka27.voting.web.controller.restaurant.UserRestaurantController.REST_URL;

public class UserRestaurantControllerTest extends AbstractControllerTest {

    @Test
    void getUnauthorized() throws Exception {
        perform(MockMvcRequestBuilders.get(REST_URL))
                .andExpect(status().isUnauthorized());
    }

    @Test
    @WithUserDetails(value = USER_MAIL)
    void getAll() throws Exception {
        perform(MockMvcRequestBuilders.get(REST_URL))
                .andExpect(status().isOk())
                .andDo(print())
                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                .andExpect(RESTAURANT_MATCHER.contentJson(restaurants));
    }
}
